# docs/swagger.yaml
basePath: /
definitions:
...
  internal_adapters_primary_api_order.getOrderResp:
    properties:
      address_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_order.orderItemInfoResp'
        type: array
      status:
        type: string
      total_amount:
        type: number
      user_id:
        type: string
    type: object
  internal_adapters_primary_api_order.listOrdersResp:
    properties:
      orders:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_order.orderInfoResp'
        type: array
    type: object
  internal_adapters_primary_api_order.orderInfoResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        type: string
      total_amount:
        type: number
    type: object
  internal_adapters_primary_api_order.orderItemInfoResp:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
      unit_price:
        type: number
    type: object
  internal_adapters_primary_api_order.orderItemReq:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    type: object
  internal_adapters_primary_api_order.placeOrderReq:
    properties:
      address_id:
        type: string
      items:
        items:
          $ref: '#/definitions/internal_adapters_primary_api_order.orderItemReq'
        type: array
    type: object
  internal_adapters_primary_api_order.placeOrderResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        type: string
      total_amount:
        type: number
    type: object
  internal_adapters_primary_api_order.updateStatusReq:
    properties:
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    A RESTful e-commerce API built with Go using hexagonal architecture.
    Features: User authentication (JWT), Product management, Order processing, Address management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /admin/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an order (admin only)
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_order.updateStatusReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update order status (Admin)
      tags:
      - Admin
...
  /orders:
    get:
      consumes:
      - application/json
      description: Get all orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_order.listOrdersResp'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List user orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Create a new order with items
      parameters:
      - description: Order data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_adapters_primary_api_order.placeOrderReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_order.placeOrderResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Place a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: Get order details by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_adapters_primary_api_order.getOrderResp'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get order details
      tags:
      - Orders
  /orders/{id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Cancel an order
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"